feat(mcp): relocate Inspect and emit recon facts
Move diagnostic orchestration out of the protocol adapter and surface final URL, status, scheme downgrade, and native alternate feeds from the same fetch.
